+++ innopoints/views/__init__.py
@@ -3,5 +3,6 @@ from .project import *
 from .activity import *
 from .file import *
 from .product import *
+from .variety import *
 
 from .views import *  # temporary, until all views have been moved
